### 2019-11-01 #1853 @whedon
[PRE REVIEW]: ytree: A Python package for analyzing merger trees
**Submitting author:** @brittonsmith (<a href="http://orcid.org/0000-0002-6804-630X">Britton Smith</a>)
**Repository:** <a href="https://github.com/ytree-project/ytree" target ="_blank">https://github.com/ytree-project/ytree</a>
**Version:** v2.2
