### 2019-11-01 #1853 @Kevin-Mattheus-Moerman
@brittonsmith Your paper has compilation issues. Please study our guidelines and see if you can trouble shoot these issues.
https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles

You can use this link to test the compilation of your paper: https://whedon.theoj.org/
