### 2019-11-01 #1853 @brittonsmith
@Kevin-Mattheus-Moerman, thanks for the note. I think the problem is that the paper is in a subdirectory of the repo linked above. The paper can be found [here](https://github.com/ytree-project/ytree/tree/master/paper). I thought that was the url I gave in the submission form. In any case, if I use that url in the compilation tester, it tells me that paper compilation is complete, although I get a "Not Found" message when I click to download the preview. Can you tell me what action I should take now?
