### 2019-11-01 #1853 @danielskatz
👋 @brittonsmith - It looks the the line after the metadata containing ```---``` is missing - please add this line to the paper.md and issue ```@whedon generate pdf``` as a new comment here to try again
