+++ paper/paper.md
@@ -21,6 +21,7 @@ affiliations:
    index: 2
 date: 4 October 2019
 bibliography: paper.bib
+---
 
 # Summary
 
