### 2019-11-05 #1853 @xuanxu
:wave: @cmbiwer would you be willing to review this paper? I think your knowledge of astronomy and Python makes you a great fit for this submission.
