### 2019-11-06 27aa96f
Fix reference formatting issues.
