+++ paper/paper.bib
@@ -109,7 +109,7 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{ctrees,
+@ARTICLE{cote2018,
        author = {{C{\^o}t{\'e}}, Benoit and {Silvia}, Devin W. and {O'Shea}, Brian W. and
          {Smith}, Britton and {Wise}, John H.},
         title = "{Validating Semi-analytic Models of High-redshift Galaxy Formation Using Radiation Hydrodynamical Simulations}",
@@ -129,7 +129,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2019ApJ...882..156H,
+@ARTICLE{hummels2019,
        author = {{Hummels}, Cameron B. and {Smith}, Britton D. and {Hopkins}, Philip F. and
          {O{\textquoteright}Shea}, Brian W. and {Silvia}, Devin W. and
          {Werk}, Jessica K. and {Lehner}, Nicolas and {Wise}, John H. and
@@ -151,7 +151,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2018MNRAS.480.3762S,
+@ARTICLE{smith2018,
        author = {{Smith}, Britton D. and {Regan}, John A. and {Downes}, Turlough P. and
          {Norman}, Michael L. and {O'Shea}, Brian W. and {Wise}, John H.},
         title = "{The growth of black holes from Population III remnants in the Renaissance simulations}",
@@ -170,7 +170,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2019MNRAS.489.4574G,
+@ARTICLE{garrisonkimmel2019,
        author = {{Garrison-Kimmel}, Shea and {Wetzel}, Andrew and {Hopkins}, Philip F. and
          {Sanderson}, Robyn and {El-Badry}, Kareem and {Graus}, Andrew and
          {Chan}, T.~K. and {Feldmann}, Robert and {Boylan-Kolchin}, Michael and
