### 2019-11-11 #1881 @whedon
[REVIEW]: ytree: A Python package for analyzing merger trees
**Submitting author:** @brittonsmith (<a href="http://orcid.org/0000-0002-6804-630X">Britton Smith</a>)
**Repository:** <a href="https://github.com/ytree-project/ytree" target ="_blank">https://github.com/ytree-project/ytree</a>
**Version:** 2.3
