### 2019-11-11 #1881 @xuanxu
@brittonsmith, @mgckind, @aureliocarnero: this is the review thread for the paper. All of our communications will happen here from now on.
