### 2019-11-15 #1881 @aureliocarnero
Hello @brittonsmith.
Your code does not contain a plain-text LICENSE file with the contents of an OSI approved software license.
Please add one to the repository in order to continue with the review.
Cheers
