### 2019-11-18 #1881 @aureliocarnero
Thank you @brittonsmith - Could you please clarify me which license from this list, https://opensource.org/licenses/alphabetical correspond to yours.? Cheers
