### 2019-11-18 #1881 @aureliocarnero
@brittonsmith I'm having troubles installing your code with pip.
It tells me the following:
  Stored in directory: /home/carnero/.cache/pip/wheels/63/9d/8e/37c3f6506ed3f152733a699e92d8e0c9f5e5f01dea262f80ad
Successfully built mpmath
Installing collected packages: mpmath, sympy, yt, ytree
  Found existing installation: sympy 0.7.6.1
ERROR: Cannot uninstall 'sympy'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall.

Any idea how can I solve this?
