### 2019-11-18 #1881 @aureliocarnero
@brittonsmith also, using the git clone, pip install instruction for the developer version, I also found this problem (maybe related)

ERROR: Exception:
Traceback (most recent call last):
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/cli/base_command.py", line 188, in main
    status = self.run(options, args)
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/commands/install.py", line 345, in run
    resolver.resolve(requirement_set)
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/legacy_resolve.py", line 196, in resolve
    self._resolve_one(requirement_set, req)
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/legacy_resolve.py", line 359, in _resolve_one
    abstract_dist = self._get_abstract_dist_for(req_to_install)
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/legacy_resolve.py", line 291, in _get_abstract_dist_for
    req, self.require_hashes, self.use_user_site, self.finder,
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/operations/prepare.py", line 255, in prepare_editable_requirement
    req.check_if_exists(use_user_site)
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_internal/req/req_install.py", line 418, in check_if_exists
    self.req.name
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_vendor/pkg_resources/__init__.py", line 481, in get_distribution
    dist = get_provider(dist)
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_vendor/pkg_resources/__init__.py", line 357, in get_provider
    return working_set.find(moduleOrReq) or require(str(moduleOrReq))[0]
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_vendor/pkg_resources/__init__.py", line 900, in require
    needed = self.resolve(parse_requirements(requirements))
  File "/home/carnero/anaconda2/lib/python2.7/site-packages/pip/_vendor/pkg_resources/__init__.py", line 791, in resolve
    raise VersionConflict(dist, req).with_context(dependent_req)
ContextualVersionConflict: (yt 3.3.5 (/home/carnero/anaconda2/lib/python2.7/site-packages), Requirement.parse('yt>=3.4'), set(['ytree']))
