### 2019-11-18 #1881 @brittonsmith
@aureliocarnero, the license is the [BSD-3-Clause](https://opensource.org/licenses/BSD-3-Clause).

For the installation issues, I'm not entirely sure. In the first instance, it looks like sympy was installed in a manner that cannot be upgraded by pip. In the second case, it looks like there is a conflicting version of yt that also cannot be updated. My recommendation would be to try creating a fresh conda environment from which to try installing from pip or source. I've just tried this now myself and it's working for me. Barring that, you might need to reinstall anaconda/miniconda. Please let me know if none of those suggestions work.
