### 2019-11-23 #1881 @aureliocarnero
Hello @brittonsmith Im trying to download the data from here: https://ytree.readthedocs.io/en/latest/Data.html trying both the manual download and the one using the girder-cl app and it failed in both cases.
Could you please check that the data is correct? Maybe it is just me, but it would be good to know.
Cheers
