### 2019-11-23 #1881 @aureliocarnero
Also, in terms of documentation, I think more information should be given in the landing page of the github. In my opinion, you do not explain well what are the application of the trees. May I suggest you explain the context is used, in terms of cosmological simulations. If it can be used for other applications than that, please explain as well
