### 2019-11-23 #1881 @aureliocarnero
You should also explain better what this mean, please:

-    loading merger-tree data from the Amiga Halo Finder, Consistent-Trees, Rockstar, LHaloTree, and treefarm formats
-    on-demand loading of trees and fields
-    symbolic units, derived fields, and alias fields
-    saving trees to a universal format

Plus you have a typo, correct the following error:  see the full ytree documenation.

Cheers
Aurelio
