### 2019-11-24 #1881 @aureliocarnero
In term of the paper, it would be good to give it a last review on the text to improve readability.
