### 2019-12-01 #1881 @aureliocarnero
So @brittonsmith please read comments above.
Some changes in the documentation front needs to be done before I give green light for the publication.
Still I need to mark:
1) A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
Please improve the landing page introduction stating clearly what the program is for.
2) Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
Please give a last reading to the paper, improving it a bit
