### 2019-12-01 #1881 @brittonsmith
@aureliocarnero, sorry I was absolutely swamped last week and also wasn't sure if I should wait for a second review. Thank you for your comments. I'm happy to clarify the text on the landing page as per all of your suggestions there. I will take care of everything this week.

As to your second comment above, I'm not sure I understand what you would like to see improved about the paper itself. Could you please clarify this?
