### 2019-12-03 #1881 @mgckind
> @mgckind: How is your review going?

Hi @xuanxu , I got it installed and started looking at the documentation  (and getting some data bbut that went too slow which I'll revisit ) and reading @aureliocarnero comments. I've been traveling more than I planned to. I'll get this done as soon as I can, thanks!
