### 2019-12-03 b78f458
Add statement of need and brief code demo to README.
