### 2019-12-03 #1881 @brittonsmith
@aureliocarnero, I've issued a [pull request](https://github.com/ytree-project/ytree/pull/34) to address your comments on the landing page. I chose to remove the bullet point list of features and to replace it with a short code demo as that seemed more illustrative.

Can you provide some more specific comments on how you'd like to see the paper improved?
