### 2019-12-03 #1881 @aureliocarnero
I answered through github I think. The README is much better, thanks. I
will send you comments about the paper shortly.

El mar., 3 dic. 2019 a las 18:17, Britton Smith (<notifications@github.com>)
escribió:

> @aureliocarnero <https://github.com/aureliocarnero>, I've issued a pull
> request <https://github.com/ytree-project/ytree/pull/34> to address your
> comments on the landing page. I chose to remove the bullet point list of
> features and to replace it with a short code demo as that seemed more
> illustrative.
>
> Can you provide some more specific comments on how you'd like to see the
> paper improved?
>
