### 2019-12-05 #1881 @aureliocarnero
@brittonsmith about improving the readability of the paper. I think it could be improve, for example, I suggest the following changes:

In the phrase:
1) In the early Universe, matter is very evenly distributed, with only small fluxuations
about the average density.
Change to
In the early Universe, matter is distributed homogeneously, with only small fluctuations
about the average density.

2) Halos grow via accretion of the surrounding material and mergers with other halos.
Change to
Halos grow via accretion of the surrounding material and by merging with other halos.

3) This process of merging to form increasingly massive halos is very naturally conceptualized as an inverted tree, where small branches connect up to continually larger ones, leading eventually to the trunk.
Change to
This process of merging to form increasingly massive halos is naturally conceptualized as an inverted tree, where small branches connect up to continually larger ones, leading eventually to a trunk.

4) One of the main products of a cosmological simulation is a series of catalogs of all halos within
the simulated volume at a number of epochs.
Change to
One of the main products of cosmological simulations is a series of catalogs of halos within
the simulated volume at different epochs.

5) Halos within succesive catalogs can be linked together to create merger trees that describe a halo’s growth history.
Change to
Halos within successive epochs can be linked together to create merger trees that describe a halo’s growth history.

6) About figure 1. It is confusing for, according to the text, first small halos are formed and then more massive halos results. According to the caption, it seems the contrary. According to the explanation in the text, descendents are going upward, not downward. Could you check this please? Also, it would be nice if on the left of the plot, you put an arrow representing time, this would facilitate the explanation

Cheers
