### 2019-12-05 #1881 @brittonsmith
@aureliocarnero, thanks for your suggestions on the paper. I've implemented them all, including for Figure 1, in [this PR](https://github.com/ytree-project/ytree/pull/35). The wording in the Figure caption was a mistake; I had ancestor and descendents switched. It should make sense now. Anyway, please let me know if that looks ok.
