### 2019-12-10 #1881 @aureliocarnero
After checking all the issues, I consider this publication is ready for submission and publication to JOSS
