### 2019-12-12 ytree-project/ytree#36 @mgckind
JOSS Review: paper
Hi,
Paper review is [here](https://github.com/openjournals/joss-reviews/issues/1881)
Excellent package and paper, I do have 2 minor suggestions:

Intro:
"A range of negative consequences arise from this situation, including difficulty of" --> "A range of negative consequences arise from this situation, including **the** difficulty of"

Figure 1: Might be obvious (but for those no familiar entirely with the nomenclature) it might worth mentioning that the sizes of the circles are proportional to the masses.
also,
"Red circles denote the line of the most massive ancestors of the primary halo" --> "Red circles denote the line of the most massive ancestors of the primary halo **at a given time step**"

"Britton acknowledges the amazing yt community" --> "Britton acknowledges the amazing `yt` community "
