### 2019-12-12 ytree-project/ytree#37 @mgckind
JOSS Review: Functionality
Hi,
Paper review is [here](https://github.com/openjournals/joss-reviews/issues/1881)

I've worked with merger trees in the past, after using a sample data and going through the documentation there are 2 issues I think are missing
- visualization of a tree example: even though this is not a visualization package, it is not clear from the paper and/or documentation how you can visualize one of these trees, like for example how to replicate Figure 1. Is a little miss leading the fact that there are no at least few copy/paste instructions on how to visualize a tree, it'd be great if this is added or at least a link where users can find the info

- A common query I remember is to find the *age* of a halo, there are different ways to define age but none of them is covered or explained here, for example, one might be the moment when the halo reached 50% of its final mass. I think I hacked a way to get this but (as a suggestion) it'd be good to at least explore these extra properties (besides the current list of properties for each node)

Thanks!
