### 2019-12-12 #1881 @mgckind
@brittonsmith , @xuanxu.  I'm sorry again for being late but here I am.
Great package and nice documentation, I went through every point, I checked aurelio's comments and installed and used `ytree`. Everything looks good to me.
 I've created 2 issues in the original repo in regards to the [paper](https://github.com/ytree-project/ytree/issues/36) and some [functionalities](https://github.com/ytree-project/ytree/issues/37). These are minor suggestions, other than that and after @aureliocarnero comments being addressed the paper looks good.  Congrats!
