### 2019-12-12 7c15c16
Paper update from JOSS review.
