### 2019-12-12 ytree-project/ytree#36 @brittonsmith
Resolved in PR #38.
