### 2019-12-14 a8ea38e
Add clear_graph decorator and docstrings.
