### 2019-12-14 06a0803
Add plot tests.
