### 2019-12-14 8c0ad23
Add plotting to api reference.
