### 2019-12-15 6fa30b5
Fix a couple plotting issues.
