### 2019-12-15 db7b0ae
Add plotting docs.
