+++ doc/source/index.rst
@@ -37,6 +37,7 @@ Table of Contents
    Data.rst
    Arbor.rst
    Fields.rst
+   Plotting.rst
    Conduct.rst
    Contributing.rst
    Developing.rst
