+++ doc/source/Loading.rst
@@ -17,9 +17,9 @@ of the first ".parameter" file.
 
 .. code-block:: python
 
-   import ytree
-   a = ytree.load("ahf_halos/snap_N64L16_000.parameter",
-                  hubble_constant=0.7)
+   >>> import ytree
+   >>> a = ytree.load("ahf_halos/snap_N64L16_000.parameter",
+   ...                hubble_constant=0.7)
 
 .. note:: Three important notes about loading AHF data:
 
@@ -46,8 +46,8 @@ format is typically one or a few files with a naming convention like
 
 .. code-block:: python
 
-   import ytree
-   a = ytree.load("consistent_trees/tree_0_0_0.dat")
+   >>> import ytree
+   >>> a = ytree.load("consistent_trees/tree_0_0_0.dat")
 
 Rockstar Catalogs
 -----------------
@@ -60,8 +60,8 @@ load in this format, simply provide the path to one of these files.
 
 .. code-block:: python
 
-   import ytree
-   a = ytree.load("rockstar/rockstar_halos/out_0.list")
+   >>> import ytree
+   >>> a = ytree.load("rockstar/rockstar_halos/out_0.list")
 
 LHaloTree
 ---------
@@ -88,8 +88,8 @@ catalog can be loaded from the path to one of the tree files.
 
 .. code-block:: python
 
-   import ytree
-   a = ytree.load("lhalotree/trees_063.0")
+   >>> import ytree
+   >>> a = ytree.load("lhalotree/trees_063.0")
 
 Both the scale factor and parameter files can be specified explicitly
 through keyword arguments if they do not match the expected pattern
@@ -97,36 +97,36 @@ or are located in a different directory than the tree files.
 
 .. code-block:: python
 
-   a = ytree.load("lhalotree/trees_063.0",
-                  parameter_file="lhalotree/param.txt",
-		  scale_factor_file="lhalotree/a_list.txt")
+   >>> a = ytree.load("lhalotree/trees_063.0",
+   ...                parameter_file="lhalotree/param.txt",
+   ...                scale_factor_file="lhalotree/a_list.txt")
 
 The scale factors and/or parameters themselves can also be passed
 explicitly from python.
 
 .. code-block:: python
 
-   import numpy as np
-   parameters = dict(HubbleParam=0.7, Omega0=0.3, OmegaLambda=0.7,
-       BoxSize=62500, PeriodicBoundariesOn=1, ComovingIntegrationOn=1,
-       UnitVelocity_in_cm_per_s=100000, UnitLength_in_cm=3.08568e21,
-       UnitMass_in_g=1.989e+43)
-   scale_factors = [ 0.0078125,  0.012346 ,  0.019608 ,  0.032258 ,  0.047811 ,
-        0.051965 ,  0.056419 ,  0.061188 ,  0.066287 ,  0.071732 ,
-        0.07754  ,  0.083725 ,  0.090306 ,  0.097296 ,  0.104713 ,
-        0.112572 ,  0.120887 ,  0.129675 ,  0.13895  ,  0.148724 ,
-        0.159012 ,  0.169824 ,  0.181174 ,  0.19307  ,  0.205521 ,
-        0.218536 ,  0.232121 ,  0.24628  ,  0.261016 ,  0.27633  ,
-        0.292223 ,  0.308691 ,  0.32573  ,  0.343332 ,  0.361489 ,
[…]
