### 2019-12-15 09497f0
Add missing functions to api reference.
