+++ doc/source/api_reference.rst
@@ -17,6 +17,9 @@ used to save the entire arbor or individual trees.
 
    ~ytree.data_structures.arbor.load
    ~ytree.data_structures.arbor.Arbor
+   ~ytree.data_structures.arbor.Arbor.add_alias_field
+   ~ytree.data_structures.arbor.Arbor.add_analysis_field
+   ~ytree.data_structures.arbor.Arbor.add_derived_field
    ~ytree.data_structures.arbor.Arbor.save_arbor
    ~ytree.data_structures.arbor.Arbor.select_halos
    ~ytree.data_structures.tree_node.TreeNode.save_tree
