### 2019-12-16 81ea897
Add examples to docs.
