+++ doc/source/index.rst
@@ -38,6 +38,7 @@ Table of Contents
    Arbor.rst
    Fields.rst
    Plotting.rst
+   Examples.rst
    Conduct.rst
    Contributing.rst
    Developing.rst
