### 2019-12-16 55a1458
Add graphviz install to travis and use python2-safe exception.
