### 2019-12-16 #1881 @brittonsmith
@mgckind, I've just issued a new [pull request](https://github.com/ytree-project/ytree/pull/39) to address your [functionality comments](https://github.com/ytree-project/ytree/issues/37). Please let me know if that looks ok.
