### 2019-12-16 #1881 @mgckind
I went through the PR and it looks awesome, thanks for the extra effort of adding those. Congrats on such a nice module!. @xuanxu , @brittonsmith This is ready to go.
