+++ doc/source/changelog.rst
@@ -13,10 +13,33 @@ The `CREDITS file
 contains the most up-to-date list of everyone who has contributed to the
 ``ytree`` source code.
 
+Version 2.3
+-----------
+
+Release date: *December 17, 2019*
+
+This release marks the `acceptance of the ytree paper
+<https://github.com/openjournals/joss-reviews/issues/1881>`__ in
+`JOSS <https://joss.theoj.org/>`__.
+
+New Features
+^^^^^^^^^^^^
+
+ * Add TreePlot for plotting and examples docs
+   (`PR #39 <https://github.com/ytree-project/ytree/pull/39>`__)
+
+Enhancements
+^^^^^^^^^^^^
+
+ * Add time field
+   (`PR #25 <https://github.com/ytree-project/ytree/pull/25>`__)
+ * Move treefarm module to separate package
+   (`PR #28 <https://github.com/ytree-project/ytree/pull/28>`__)
+
 Version 2.2.1
 -------------
 
-Release data: *October 24, 2018*
+Release date: *October 24, 2018*
 
 Enhancements
 ^^^^^^^^^^^^
