### 2019-12-17 c341026
Add note about dropping python 2 support after release.
