### 2019-12-17 #1881 @brittonsmith
Hi @xuanxu, I've just done the release.

Version: 2.3
DOI: [10.5281/zenodo.3580978](https://doi.org/10.5281/zenodo.3580978)

The metadata for the Zenodo entry looks good to me. Our ORCIDs are there, too. I think we're ready to go.

Thanks for all your help!
