### 2019-12-17 #1881 @xuanxu
Thanks @brittonsmith!
Everything is ready :tada:
@openjournals/joss-eics sending it your way for final acceptance! :rocket:
